# swagger.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "NeuroSuitUp API",
        "description": "Tools for communicating with the NeuroSuitUp database",
        "version": "1.0.0",
        "contact": {
        "name": "Apostolos Praftsiotis",
        "email": "apospraf@gmail.com"
        }
    },
    "servers": [
        {
            "url":"http://localhost:3000",
            "description":"Temp localhost server"
        },
        {
            "url": "https://neurosuituptest.herokuapp.com/",
            "description": "Heroku free server"
        }, 
        {
            "url": "http://192.168.0.188:3000",
            "description": "Heroku free server"
        }
    ],
    "components": {
      "schemas": { 
...
        "PoseInsert": {
            "type": "object",
            "properties": {
                "codename": {
                    "type": "string",
                    "description": "The codename of the specific pose"
                },
                "isPublic": {
                    "type": "boolean",
                    "description": "Whether or not the specific pose is available to every instructor"
                },
                "angles": {
                    "type": "object",
                    "description": "The 16 angles, that describe the specific pose",
                    "properties": {
                        "SyL": {
                            "type": "string",
                            "description": "Shoulder y angle of left hand"
                        },
                        "SzL": {
                            "type": "string",
                            "description": "Shoulder z angle of left hand"
                        },
                        "UxL": {
                            "type": "string",
                            "description": "Upper arm x angle of left hand"
                        },
                        "UyL": {
                            "type": "string",
                            "description": "Upper arm y angle of left hand"
                        },
                        "UzL": {
                            "type": "string",
                            "description": "Upper arm z angle of left hand"
                        },
                        "FyL": {
                            "type": "string",
                            "description": "Forearm y angle of left hand"
                        },
                        "FzL": {
                            "type": "string",
                            "description": "Forearm z angle of left hand"
                        },
                        "HxL": {
                            "type": "string",
                            "description": "Hand x angle of left hand"
                        },
                        "HzL": {
                            "type": "string",
                            "description": "Hand z angle of left hand"
                        },
                        "SyR": {
                            "type": "string",
                            "description": "Shoulder y angle of right hand"
                        },
                        "SzR": {
                            "type": "string",
                            "description": "Shoulder z angle of right hand"
                        },
                        "UxR": {
                            "type": "string",
                            "description": "Upper arm x angle of right hand"
                        },
                        "UyR": {
                            "type": "string",
                            "description": "Upper arm y angle of right hand"
                        },
                        "UzR": {
                            "type": "string",
                            "description": "Upper arm z angle of right hand"
                        },
                        "FyR": {
                            "type": "string",
                            "description": "Forearm y angle of right hand"
                        },
                        "FzR": {
                            "type": "string",
                            "description": "Forearm z angle of right hand"
                        },
                        "HxR": {
                            "type": "string",
                            "description": "Hand x angle of right hand"
                        },
                        "HzR": {
                            "type": "string",
                            "description": "Hand z angle of right hand"
                        }
                    }
                }
            }, 
            "required": ["codename", "isPublic", "angles"]
        },
        "Pose": {
            "allOf": [
                {
                    "$ref": "#/components/schemas/PoseInsert"
                },
                {
                    "type": "object",
                    "properties": {
                        "_id": {
                            "type": "string",
                            "description": "The pose id (ObjectId mongo)"
                        }
                    } 
                }
            ]
        },
        "TaskInsert": {
            "type": "object",
            "properties": {
                "codename": {
                    "type": "string",
                    "description": "The codename of the specific task"
                },
                "isPublic": {
                    "type": "boolean",
                    "description": "Whether or not the specific task is available to every instructor"
                },
                "difficulty": {
                    "type": "integer",
                    "description": "The difficulty of the specific task (1 - 10)"
                },
                "poses": {
                    "type": "array",
                    "description": "Array containing the poses that define the specific task",
                    "items": {
                        "type": "string",
                        "description": "The pose id (ObjectId mongo)"
                    }
                }
            },
            "required": ["codename", "isPublic", "difficulty", "poses"]
        },
        "Task": {
          "allOf": [
              {"$ref": "#/components/schemas/TaskInsert"},
              {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "description": "The task id (ObjectId, mongodb)"
                    }
                }
              }
          ]  
        },
...
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }, 
        "Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
      },
      "parameters": {
          "selectstring": {
              "in": "header",
              "name": "selectstring",
              "description": "Selects the fields of the returned resource",
              "schema": {
                  "type": "string"
              },
              "allowEmptyValue": true
          }
      },
      "responses": {
...
        "NotFound": {
            "description": "The specified resource was not found",
            "content": {
                "application/json": {
                    "schema": {
                        "$ref": "#/components/schemas/Error" 
                    }
                }
            }
        }, 
        "Unauthorized": {
            "description": "Authorization failed, either by unsuccessfull verification of token, or by limited permissions",
            "content": {
                "application/json": {
                    "schema": {
                        "$ref": "#/components/schemas/Error"
                    }
                }
            }
        }, 
        "InternalError": {
            "description": "Internal or structural error",
            "content": {
                "application/json": {
                    "schema": {
                        "$ref": "#/components/schemas/Error"
                    }
                }
            }
        }
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    },
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "paths": {
...
      "/poses/codenameExists/{codename}": {
          "get": {
            "summary": "Check if codename of pose exists in the database",
            "tags": ["poses"],
            "parameters": [{
                "in": "path",
                "name": "codename",
                "description": "The codename to be searched",
                "required": true,
                "schema": {
                    "type": "string"
                }
            }],
            "responses": {
                "200": {
                    "description": "Query was successful",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type":"boolean"
                            }
                        }
                    }
                }, 
                "401": {
                    "$ref": "#/components/responses/Unauthorized"
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
          }
      },
      "/poses": {
          "get": {
              "summary": "Get all available poses for the logged in instructor",
              "tags": ["poses"],
              "parameters": [{
                  "$ref": "#/components/parameters/selectstring"
              }],
              "responses": {
                  "200": {
                      "description": "Successfully fetched all available poses",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "type": "object",
                                  "properties": {
                                      "count": {
                                          "type": "integer",
                                          "description": "The number of poses"
                                      }, 
                                      "poses": {
                                          "type": "array",
                                          "description": "Array containing all the poses",
                                          "items": {
                                              "$ref": "#/components/schemas/Pose"
                                          }
                                      }
                                  }
                              }
                          }
                      }
                  },
                  "401": {
                      "$ref": "#/components/responses/Unauthorized"
                  },
                  "500": {
                      "$ref": "#/components/responses/InternalError"
                  }
              }
          },
          "post": {
            "summary": "Insert a new pose to the database",
            "tags": ["poses"],
            "parameters": [{
                "$ref": "#/components/parameters/selectstring"
            }],
            "requestBody": {
                "description": "The pose to be added",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "$ref": "#/components/schemas/PoseInsert"
                        }
                    }
                }
            }, 
            "responses": {
                "201": {
                    "description": "Successfully inserted pose to database",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "The new pose id",
                                "properties": {
                                    "_id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "401": {
                    "$ref": "#/components/responses/Unauthorized"
                },
                "409": {
                    "description": "Pose codename already exists",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
          },
          "patch": {
              "summary": "Update pose angles with a given codename",
              "tags": ["poses"],
              "parameters": [
                  {
                      "in": "path",
                      "name": "codename",
                      "description": "Codename of the requested pose",
                      "required": true,
                      "schema": {
                        "type": "string"
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successfully update pose",
                      "content": {
                          "application/json": {
                              "schema": {
                                "$ref": "#/components/schemas/Message"
                              }
                          }
                      }
                  },
                  "401": {
                      "$ref": "#/components/responses/Unauthorized"
                  },
                  "500": {
                      "$ref": "#/components/responses/InternalError"
                  }
              }
          }
      }, 
      "/poses/{poseId}": {
        "parameters": [{
            "in": "path",
            "name": "poseId",
            "description": "The pose id",
            "required": true,
            "schema": {
                "type": "string"
            }
        }],
        "get": {
            "summary": "Get pose by id",
            "tags": ["poses"],
            "parameters": [{
                "$ref": "#/components/parameters/selectstring"
            }],
            "responses": {
                "200": {
                    "description": "Pose succefully found",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Pose"
                            }
                        }
                    }
                },
                "401": {
                    "$ref": "#/components/responses/Unauthorized"
                }, 
                "404": {
                    "description": "Pose not found",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
        },
        "delete": {
            "summary": "Delete a pose by id",
            "tags": ["poses"],
            "responses": {
                "200": {
                    "description": "Successfully deleted pose",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "401": {
                    "$ref": "#/components/responses/Unauthorized"
                },
                "403": {
                    "description": "Pose is not owned, thus cannot be deleted",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "409": {
                    "description": "Some tasks will be affected after deleting this pose",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "message": {
                                        "type": "string"
                                    },
                                    "tasks": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Task"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
        }
      },
      "/poses/fetchPoses": {
          "post": {
              "summary": "Fetch specific poses with ids, located at the body of the request",
              "tags": ["poses"],
              "parameters": [
                  {
                      "$ref": "#/components/parameters/selectstring"
                  }
              ],
              "requestBody": {
                  "description": "The array of the requested pose ids",
                  "content": {
                      "application/json": {
                          "schema": {
                              "type": "array",
                              "items": {
                                  "type": "string",
                                  "description": "The pose id (ObjectId mongo)"
                              }
                          },
                          "example": "{['615c94503b3ffd5cccec496a', '615c945e3b3ffd5cccec496b']}"
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": "Successfully fetched all poses",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "type": "object",
                                  "properties": {
                                      "count": {
                                          "type": "integer",
                                          "description": "The number of found poses"
                                      },
                                      "poses": {
                                          "type": "array",
                                          "description": "Array containg the pose ids",
                                          "items": {
                                              "type": "string"
                                            }
                                      }
                                  }
                              }
                          }
                      }
                  },
                  "401": {
                      "$ref": "#/components/responses/Unauthorized"
                  },
                  "500": {
                      "$ref": "#/components/responses/InternalError"
                  }
              }
          }
      },
      "/poses/copyPose/{poseId}": {
        "patch": {
            "tags": ["poses"],
            "summary": "Copy pose from public space to user space",
            "parameters": [{
                "in": "path",
                "name": "poseId",
                "required": true,
                "schema": {
                    "type": "string",
                    "description": "The id of the pose to be copied"
                }
            }],
            "responses": {
                "201": {
                    "description": "Pose was successufully copied",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "401": {
                    "$ref": "#/components/responses/Unauthorized"
                },
                "403": {
                    "description": "Pose is not owned, thus cannot be copied",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "404": {
                    "$ref": "#/components/responses/NotFound"
                },
                "409": {
                    "description": "Pose already owned",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
        }
      },
      "/poses/{codename}": {
        "patch": {
            "tags": ["poses"],
            "summary": "Edit pose, only accessible by pose owner",
            "parameters": [{
                "in": "path",
                "name": "codename",
                "description": "The codename of the pose",
                "required": true,
                "schema": {
                    "type": "string"
                }
            }],
            "responses": {
                "200": {
                    "description": "Successfully updated pose",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                }, 
                "401": {
                    "$ref": "#/components/responses/Unauthorized"
                },
                "403": {
                    "description": "Pose is not owned, thus cannot be edited",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "404": {
                    "$ref": "#/components/responses/NotFound"
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
        }
      },
...
    }
}